PCX-1020 handles the error cases in the JSON.parse
