+++ demo/src/pages/index.js
@@ -1,17 +1,9 @@
 import Products from "./Products";
 import Summary from "./Summary";
 import Thankyou from "./Thankyou";
+import { disableErrorOverlay } from "./disableErrorOverlay";
 
-window.addEventListener("message", event => {
-    if (JSON.parse(event?.data)?.__postRobot__?.data?.name === "onError") {
-        const interval = setInterval(() => {
-            const element = document.getElementById("react-refresh-overlay");
-            if (element) {
-                clearInterval(interval);
-                element.remove();
-            }
-        }, 0);
-    }
-});
+// removes the overlay set by the nwb
+disableErrorOverlay();
 
 export { Products, Summary, Thankyou };
